+++ src/components/history/HistoryInfo.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="main">
-    <h5> Record of Service</h5>
+    <h5> Service History</h5>
     <div v-if="message" class="alert-success"/>
     <div class="align-content-center">
       <table class="table">
@@ -18,7 +18,7 @@
         </thead>
         <tbody>
         <tr v-for="(history, index) in histories" v-bind:key="index">
-          <td>{{index + 1}}</td>
+          <td>{{ index + 1 }}</td>
           <td>{{ history.mda.name }}</td>
           <td>{{ history.department.name }}</td>
           <td>{{ history.designation.name }}</td>
@@ -36,7 +36,7 @@
         </tr>
         </tbody>
       </table>
-      <div  >
+      <div>
         <router-link :to="{name:'addHistory'}" class="btn btn-success ">Add</router-link>
         <router-link :to="{name:'promotionInfo'}"
                      class="i btn btn-success" v-if="histories.length">Next
@@ -64,9 +64,24 @@ export default {
       });
     },
     deleteHistory(id) {
-      HistoryService.deleteHistoryById(id).then(() => {
-        this.histories = this.histories.filter(history => history.id !== id);
-      });
+      this.$dialog.confirm("If you delete this record, it'll be gone forever.", {
+        loader: true
+      })
+          .then((dialog) => {
+
+            HistoryService.deleteHistoryById(id).then(() => {
+              this.histories = this.histories.filter(history => history.id !== id);
+            });
+            setTimeout(() => {
+              console.log('Delete action completed ');
+              dialog.close();
+            }, 2500);
+          })
+          .catch(() => {
+            // Triggered when cancel button is clicked
+
+            console.log('Delete aborted');
+          });
     }
   },
   created() {
@@ -82,7 +97,7 @@ export default {
   padding-right: 30px;
 }
 
-.i{
+.i {
   margin-left: 150px;
 }
 </style>
